+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 0.0.10
 
 - Added max_concurrency setting to model.
+- Added model info endpoint.